update connection informing style stable
