+++ io/server.js
@@ -1,0 +1,27 @@
+const app = require(require.resolve('express'))();
+const server = require(require.resolve('http')).Server(app);
+const checkUserInfo = require('./checkUserInfo');
+
+const startTimestamp = new Date().getTime();
+const origins = ['enecuum.com:*'];
+
+const io = require('socket.io')(server, {
+  path: '/io',
+  serveClient: false,
+  pingInterval: 1000,
+  pingTimeout: 1000,
+  cookie: true
+});
+
+io.origins[origins];
+
+io.on('connection', (client) => {
+  client.emit('connectServer', startTimestamp);
+  checkUserInfo(client, {cookies: client.handshake.headers.cookie});
+});
+
+module.exports = {
+  io,
+  app,
+  server
+}
