+++ plugins/socket.io.js
@@ -1,6 +1,6 @@
 import io from 'socket.io-client';
 
-const socket = io((process.env.dev ? 'http' : 'https') + '://enecuum.com', {
+const socket = io((process.env.dev ? 'http' : 'https') + '://enecuum.com'+(process.env.dev ? ':8081': ''), {
   path: '/io',
   reconnectionAttempts: 30,
   reconnectionDelay: 3000
